Move venstar base entity to separate module
