+++ homeassistant/components/venstar/__init__.py
@@ -13,9 +13,7 @@ from homeassistant.const import (
     CONF_USERNAME,
     Platform,
 )
-from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
+from homeassistant.core import HomeAssistant
 
 from .const import DOMAIN, VENSTAR_TIMEOUT
 from .coordinator import VenstarDataUpdateCoordinator
@@ -59,36 +57,3 @@ async def async_unload_entry(hass: HomeAssistant, config: ConfigEntry) -> bool:
     if unload_ok:
         hass.data[DOMAIN].pop(config.entry_id)
     return unload_ok
-
-
-class VenstarEntity(CoordinatorEntity[VenstarDataUpdateCoordinator]):
-    """Representation of a Venstar entity."""
-
-    _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        venstar_data_coordinator: VenstarDataUpdateCoordinator,
-        config: ConfigEntry,
-    ) -> None:
-        """Initialize the data object."""
-        super().__init__(venstar_data_coordinator)
-        self._config = config
-        self._client = venstar_data_coordinator.client
-
-    @callback
-    def _handle_coordinator_update(self) -> None:
-        """Handle updated data from the coordinator."""
-        self.async_write_ha_state()
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return the device information for this entity."""
-        fw_ver_major, fw_ver_minor = self._client.get_firmware_ver()
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._config.entry_id)},
-            name=self._client.name,
-            manufacturer="Venstar",
-            model=f"{self._client.model}-{self._client.get_type()}",
-            sw_version=f"{fw_ver_major}.{fw_ver_minor}",
-        )
