+++ homeassistant/components/venstar/binary_sensor.py
@@ -8,8 +8,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import VenstarEntity
 from .const import DOMAIN
+from .entity import VenstarEntity
 
 
 async def async_setup_entry(
